requirements.txt
python-telegram-bot==21.9
python-pptx==0.6.22
python-docx==1.1.2
